REPO-3160: Add the email TAS tests to our builds
REPO-3160: Add the email TAS tests to our builds 

Add jolokia to the test docker image. This will be used by tas-email-tests in order to make the configurations to the server.
Enable imap server.
Expose an unused port for email server that will be used by tas for smtp settings (1126).
Expose port 1125 that will be used by the tests to change the port in real time.
Enable quota that is required for some tests.