REPO-5361/REPO-5358 Merge latest changes / Deployment modules should extend each other
In order to support the external deployment of the default keystore (MNT-21731/SEARCH-1915),
it was removed from alfresco-repository.jar and had be placed in deployment module of acs-packaging.
To support this in the new project structure, it was needed to move the keystore back to
alfresco-community-repo's deployment module and to have alfresco-enterprise-repo and acs-packaging
extend it in their own deployment modules. This avoids duplication of licenses too.