Add script to enable changing between versions of acs projects.
Note that this script will only be present in versions after this commit.  Unfortunately this
means that if you change to a release older than this commit then it will be slightly awkward
to get back again. (Something like git checkout origin/master and then rerunning the script
will work though).