Grep! [release][skip tests][skip docker_release][skip downstream] 7.2.0-A19.6
Trim line ending warnings out of the log [skip ci]

We ended up with +23,000 warning messages about the wrong line endings because of update of the copyright year in RM files and the Travis logs filled up and terminated. Greps added to remove them, while we look for a better solution.