Feature/acs 4156 fix instabilities of elasticsearch reindexing tests (#2349)
Move try with resources to extend container livetime to the length of retrieving response from query.
Increase timeout to 60 seconds. This timeout is only important in case of failures from response, success ends processing immediately, failure retries up to the timeout. If the timeout is reached, the test is failed. So extending that to 60 seconds seems better than having to rerun the test job in case of things that can take time (like reindexing in particular)